requirements.txt
fastapi>=0.100.0
uvicorn[standard]
pylatexenc
python-multipart
pydantic>=2.0
google-generativeai>=0.4.0
PyMuPDF>=1.23.0
python-docx>=1.0.0
httpx>=0.27.0
